+++ internal/functions/rendered_template.go
