+++ internal/functions/stencil.go
